move the conversation screen state
